+++ src/Elasticsearch/Endpoints/Cluster/AllocationExplain.php
@@ -59,6 +59,6 @@ class AllocationExplain extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }