+++ src/Elasticsearch/Endpoints/Explain.php
@@ -99,6 +99,6 @@ class Explain extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }