+++ src/Elasticsearch/Endpoints/Indices/ClearCache.php
@@ -59,6 +59,6 @@ class ClearCache extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }