+++ src/Elasticsearch/Endpoints/Indices/ValidateQuery.php
@@ -74,6 +74,6 @@ class ValidateQuery extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }