+++ src/Elasticsearch/Endpoints/MTermVectors.php
@@ -67,6 +67,6 @@ class MTermVectors extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'POST';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }