+++ src/Elasticsearch/Endpoints/Msearch.php
@@ -104,6 +104,6 @@ class Msearch extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }