+++ src/Elasticsearch/Endpoints/Scroll.php
@@ -100,6 +100,6 @@ class Scroll extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }