+++ src/Elasticsearch/Endpoints/Search.php
@@ -115,6 +115,6 @@ class Search extends AbstractEndpoint
      */
     public function getMethod()
     {
-        return 'GET';
+        return isset($this->body) ? 'POST' : 'GET';
     }
 }